Add missing dependeny, Fix output, Add some usage info.
Add a message the prompts the user to setup the git config if they already haven't. Only looking at the git config, disabled the environment variable configuration for the time being. Update the read me to reflect this. Fixed an issue that would cause projects title to be printed when there where only tasks that wouldn't be shown anyway because they are not started or unstarted. Added ansi color dependency.